Completes first part of migration
