+++ models/city.go
@@ -3,10 +3,7 @@ package models
 import(
   // "fmt"
   "os"
-  "math"
   s "strings"
-  "time"
-
   "github.com/gocarina/gocsv"
 )
 
@@ -23,88 +20,6 @@ type Coordinate struct {
 }
 
 var citiesDB = make(map[string]Coordinate)
-
-// Boeing 777-300ER Engine
-// Max Level Speed (at altitude)575 mph (930 km/h) at 35,000 ft (10,675 m), Mach 0.87
-const Kps = 0.2583333333 // kilometer per second assuming max speed of (930 km/h)
-
-//:::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::
-//:::                                                                         :::
-//:::  This routine calculates the distance between two points (given the     :::
-//:::  latitude/longitude of those points). It is being used to calculate     :::
-//:::  the distance between two locations using GeoDataSource (TM) prodducts  :::
-//:::                                                                         :::
-//:::  Definitions:                                                           :::
-//:::    South latitudes are negative, east longitudes are positive           :::
-//:::                                                                         :::
-//:::  Passed to function:                                                    :::
-//:::    lat1, lon1 = Latitude and Longitude of point 1 (in decimal degrees)  :::
-//:::    lat2, lon2 = Latitude and Longitude of point 2 (in decimal degrees)  :::
-//:::    unit = the unit you desire for results                               :::
-//:::           where: 'M' is statute miles (default)                         :::
-//:::                  'K' is kilometers                                      :::
-//:::                  'N' is nautical miles                                  :::
-//:::                                                                         :::
-//:::  Worldwide cities and other features databases with latitude longitude  :::
-//:::  are available at https://www.geodatasource.com                         :::
-//:::                                                                         :::
-//:::  For enquiries, please contact sales@geodatasource.com                  :::
-//:::                                                                         :::
-//:::  Official Web site: https://www.geodatasource.com                       :::
-//:::                                                                         :::
-//:::               GeoDataSource.com (C) All Rights Reserved 2018            :::
-//:::                                                                         :::
-//:::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::
-
-
-func GetRealDistance(lat1 float64, lng1 float64, lat2 float64, lng2 float64, unit ...string) float64 {
-	const PI float64 = 3.141592653589793
-
-	radlat1 := float64(PI * lat1 / 180)
-	radlat2 := float64(PI * lat2 / 180)
-
-	theta := float64(lng1 - lng2)
-	radtheta := float64(PI * theta / 180)
-
-	dist := math.Sin(radlat1) * math.Sin(radlat2) + math.Cos(radlat1) * math.Cos(radlat2) * math.Cos(radtheta)
-
-	if dist > 1 {
-		dist = 1
-	}
-
-	dist = math.Acos(dist)
-	dist = dist * 180 / PI
-	dist = dist * 60 * 1.1515
-
-	if len(unit) > 0 {
-		if unit[0] == "K" {
-			dist = dist * 1.609344
-		} else if unit[0] == "N" {
-			dist = dist * 0.8684
-		}
-	}
-	return dist
-}
-
-func GetTime(t1, t2 time.Time) (float64) {
-  diff := t2.Sub(t1)
-  return diff.Seconds()
-}
-
-// Math for calculating if transaction is pbysically (commenrcially, not military) possible
-// Assuming the maximum distance somone can travel (in KM) k/s * s
-func getPotentialDistance(kps float64, deltaSeconds float64) float64{
-  return(kps*deltaSeconds)
-}
-
-// real distance should always be = or <= to potential
-// distance (ideal distance that you can travel with ideal conditions)
-func isPosible(realDistance, potentialDistance float64) bool {
-  if (realDistance > potentialDistance) {
-    return false
-  }
-  return true
-}
 
 func init() {
   citiesFile, err := os.OpenFile("worldcities.csv", os.O_RDWR|os.O_CREATE, os.ModePerm)
